+++ src/OpenLoco/Ptr.h
@@ -1,3 +1,3 @@
 #pragma once
 #define ToPtr(type, x) ((type*)(uintptr_t)x)
-#define ToInt(x) ((int32_t)(intptr_t)x)
+#define ToInt(x) ((uintptr_t)x)
